requirements.txt
# Kafka Data Pipeline Dependencies

# Kafka client
kafka-python==2.0.2

# MongoDB driver
pymongo==4.5.0

# Streamlit for dashboard
streamlit==1.28.1

# Data manipulation and analysis
pandas==2.1.3
numpy==1.25.2

# Visualization
plotly==5.17.0

# Date and time handling
python-dateutil==2.8.2

# Logging and utilities
colorlog==6.7.0

# Environment variables
python-dotenv==1.0.0

# JSON handling (built-in, but listed for clarity)
# json - built-in module

# Threading (built-in, but listed for clarity)
# threading - built-in module

# Collections (built-in, but listed for clarity)
# collections - built-in module

# Datetime (built-in, but listed for clarity)
# datetime - built-in module

# Random (built-in, but listed for clarity)
# random - built-in module

# Time (built-in, but listed for clarity)
# time - built-in module

# Logging (built-in, but listed for clarity)
# logging - built-in module 